+++ platformio/commands/run.py
@@ -134,10 +134,11 @@ class EnvironmentProcessor(object):
                      "debug_extra_cmds", "debug_server", "debug_init_break",
                      "debug_load_cmd")
 
-    IGNORE_BUILD_OPTIONS = ("test_transport", "test_ignore", "test_port",
-                            "debug_tool", "debug_port", "debug_init_cmds",
-                            "debug_extra_cmds", "debug_server",
-                            "debug_init_break", "debug_load_cmd")
+    IGNORE_BUILD_OPTIONS = ("test_transport", "test_filter", "test_ignore",
+                            "test_port", "debug_tool", "debug_port",
+                            "debug_init_cmds", "debug_extra_cmds",
+                            "debug_server", "debug_init_break",
+                            "debug_load_cmd")
 
     REMAPED_OPTIONS = {"framework": "pioframework", "platform": "pioplatform"}